Bug fixes & refactor
- Fix regression in browser sharing
- Fix regression in guest hiding save+load buttons
- Finish moving liveshare init into its own function
