+++ src/session.ts
@@ -207,7 +207,7 @@ export async function showBrowser(url: string, title: string, viewer: string | b
                 enableScripts: true,
                 retainContextWhenHidden: true,
             });
-        if (await isHost() && config().get<boolean>('liveShare.autoShareBrowser')) {
+        if (await isHost()) {
             await shareBrowser(url, title);
         }
         panel.onDidChangeViewState((e: WebviewPanelOnDidChangeViewStateEvent) => {
@@ -226,7 +226,7 @@ export async function showBrowser(url: string, title: string, viewer: string | b
             activeBrowserPanel = undefined;
             activeBrowserUri = undefined;
             activeBrowserExternalUri = undefined;
-            if (await isHost() && config().get<boolean>('liveShare.autoShareBrowser')) {
+            if (await isHost()) {
                 closeBrowser(url);
             }
             void commands.executeCommand('setContext', 'r.browser.active', false);
